Removed drills specific method to drill.ts, made list-component more versatile
