+++ src/pages/drills/drills.ts
@@ -1,8 +1,11 @@
 import { Component } from '@angular/core';
-import { IonicPage, NavController, NavParams } from 'ionic-angular';
+import { IonicPage, NavController, NavParams, AlertController } from 'ionic-angular';
 
 //Components
 import { ListComponent } from '../components/login.component';
+
+// Firebase
+import { AngularFire, FirebaseListObservable } from 'angularfire2';
 
 
 
@@ -15,9 +18,118 @@ import { ListComponent } from '../components/login.component';
 
 
 export class Drills {
+  private drills: FirebaseListObservable<any>;
+  private uid: string;
+
   constructor(
   	public navCtrl: NavController,
-  	public navParams: NavParams) {}
+  	public navParams: NavParams,
+    public alertCtrl: AlertController,
+    public af: AngularFire
+  ) {
+    this.af.auth.subscribe(auth => {
+      if ( auth ) {
+        this.uid = auth.uid;
+        this.drills = af.database.list(this.uid+'/drills/');
+      }
+    });
+  } //constructor
+
+
+
+  public editDrill(drillId:string, drillTitle:string, drillType:string, drillLink:string) {
+    let prompt = this.alertCtrl.create({
+        title: 'Drill Name',
+        message: "Update the name for this drill",
+        inputs: [
+          {
+            name: 'title',
+            placeholder: 'Title',
+            value: drillTitle
+          },
+          {
+            name: 'type',
+            placeholder: 'Type',
+            value: drillType
+          },
+          {
+            name: 'link',
+            placeholder: 'Link',
+            value: drillLink
+          }
+        ],
+        buttons: [
+          {
+            text: 'Cancel',
+            handler: data => {
+              console.log('Cancel clicked');
+            }
+          },
+          {
+            text: 'Save',
+            handler: data => {
+              this.drills.update(drillId, {
+                title: data.title,
+                type: data.type,
+                link: data.link
+              });
+            }
+          }
+        ]
+    });
+    prompt.present();
+  } //editDrill
+
+
+
+  public addDrill() {
+    let prompt = this.alertCtrl.create({
+      title: 'Drill Name',
+      message: "Enter a name for this new song you're so keen on adding",
+      inputs: [
+        {
+          name: 'title',
+          placeholder: 'Title',
+        },
+        {
+          type: 'type',
+          placeholder: 'Type'
+        },
+        {
+          type: 'link',
+          placeholder: 'Link'
+        }
+      ],
+      buttons: [
+        {
+          text: 'Cancel',
+          handler: data => {
+            console.log('Cancel clicked');
+          }
+        },
+        {
+          text: 'Save',
+          handler: data => {
+            console.log('data - ',data);
+            this.drills.push({
+              title: data.title,
+              type: data[1],
+              link: data[2]
+            });
+          }
+        }
+      ]
+    });
+    prompt.present();
+  } //addDrill
+
+
+
+  public removeDrill(drillId: string){
+      this.drills.remove(drillId);
+  } //removeDrill
+
+
 
   ionViewDidLoad() {
     console.log('ionViewDidLoad Drills');
